Update README.md
Denji is a destitute young man who lives in a world filled with bloodthirsty monsters called Devils. One day, he befriends Pochita, who is a Devil dog-thing that is also a literal chainsaw. If that wasn't already a weird, things take a real turn when Denji is murdered by gangsters, and Pochita fuses with his soul to save his life. This turns the perpetually hungry and horny Denji into the titular Chainsaw Man, who ends up hunting other Devils for the enigmatic Makima, a woman who works for a governmental bureau of Devil Hunters, and strings Denji along with vague promises of sex and fortune, even as she treats him as little more than a dog himself.